Add s3 retry config and refactor s3_client creation as util
