+++ config/global_config.json
@@ -2,7 +2,11 @@
   "s3_bucket": {
     "uri": "dataclinic-gtfs-rt",
     "public_key": "",
-    "secret_key": ""
+    "secret_key": "",
+    "retries": {
+      "max_attempts": 3,
+      "mode": "standard"
+   }
   },
   "mobilitydatabase": {
     "url": "https://api.mobilitydatabase.org/v1",
